+++ source/chitra/elements/package.d
@@ -16,7 +16,8 @@ const ELEMENTS = [
     "rect",
     "rotate",
     "scale",
-    "translate"
+    "translate",
+    "image"
 ];
 
 public