+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Added support for rendering PNG images (`image(PATH, x, y)`).
+- Added functions to get image dimentions, `imageWidth(PATH)`, `imageHeight(PATH)`
+  and `imageSize(PATH)`. `imageSize` returns the `struct{double width, double height}`. 
 
 ## [v0.1.1] - 2024-12-04